# arc: add use of optimized versions of string functions for ARCv3 32-bit