Add ability to use optimized versions of string functions for ARCv3 32-bit CPUs with the UCLIBC_HAS_STRING_ARCH_OPT option. Add optimized **memcpy/memset/memcmp** for ARCv3 CPUs from the newlib and adapt for ARCv3 existed optimized for ARCv2 **strchr/strcmp/strcpy/strlen**.

Details:
**memset** and **memcpy**: implementations for the ```__ARC64_ARCH32__``` have been moved from the newlibc and conditional compilation for the ```__ARC_LL64__``` has been added.

**memcmp**:  has been adopted for ARCv3 32bit, copy chunks reduced form 32-byte to 16-byte; aggressive optimization that allows to read data for a next iteration in delay slots and read beyond the "count" has been  removed, as uclibc restricts such behavior and contains tests that check this.
```
nsim HS5x

timemembech with generic funcs
 standard memcpy                                      :     26.4 MB/s
 standard memset                                      :     86.4 MB/s

timemembech with optim funcs
 standard memcpy                                      :     95.0 MB/s
 standard memset                                      :    158.4 MB/s

cmptest 1Mb  with generic funcs
 strcmp time:    2.1 s
 memcmp time:    1.0 s
   
cmptest 1Mb with optim funcs
 strcmp time:    0.7 s
 memcmp time:    0.4 s

```