@shahab-vahedi 2022-09-30
Great job @pavelvkozlov ! Specially for taking care of the correct amount of reads. In Linux we do care for that, but for baremetal (newlib) we are more relaxed. One question though: Are all the new codes tested for correct functionality?