@pavelvkozlov 2022-09-30
> Are all the new codes tested for correct functionality?

I've ran all tests from standard uclibc-test pack. Test package contains a good set of tests for string functions.
```
Total skipped: 6
Total failed: 2
Total passed: 473
```
Also:
tinymembech, and my simple time measurement test.