@pavelvkozlov 2022-10-07
Thanks for the review. Then I'll prepare a patch and send it to the uclibc upstream.