@pavelvkozlov 2022-10-14
The patch has been added to the uclibc-ng upstream: [commit](https://cgit.uclibc-ng.org/cgi/cgit/uclibc-ng.git/commit/?id=663b8a0497c40a20668258bd69db13924c569c41)